Adds 64-bit reading and writing across the whole conversion path.

std gains int64, uint64 and float64, reading and writing four registers with the
same low_first and swap_bytes arguments as the 32-bit family, which moved onto
the same width-generic register helpers in ConverterTools. std.debug reports a
four-register interpretation so a 64-bit device register can be identified
during commissioning, and a converter can declare the register count it expects
so a mismatched count: is reported while the config is read.

MqttValue carries INT64, UINT64 and FLOAT64 alongside its existing types, so a
value up to UINT64_MAX is held and published exactly, and getLongDouble() parses
a command payload at full width rather than through strtod.

exprconv evaluates expressions as long double and offers int64, uint64 and flt64
register helpers, each with a bs variant, for reading and for write_as. The
integer helpers are registered only where a long double has at least 64 mantissa
bits, so on 32-bit ARM a configuration naming one is refused at startup rather
than publishing a rounded value.

Converter plugins must now export converter_plugin_abi_version. MqttValue
crosses the plugin boundary by value and its layout changed, so a plugin built
against the older headers is refused with an explanation instead of corrupting
memory. Third-party plugins have to be rebuilt.

Verified on the ci_tests branch across amd64, i386, arm64, arm/v7 and arm/v6.

+++ libmodmqttsrv/default_command_converter.cpp
@@ -38,7 +38,7 @@ DefaultCommandConverter::parseAsJson(ModbusRegisters& ret, const std::string jso
     if (!doc.IsArray())
         throw ConvException("Only json array is supported when converting to multiple registers");
 
-    if (doc.Size() != registerCount) {
+    if (static_cast<int>(doc.Size()) != registerCount) {
         throw ConvException(std::string("Wrong json array size (" + std::to_string(doc.Size()) + "), need " + std::to_string(registerCount)));
     }
